requirements.txt
PyQt6==6.7.1
SQLAlchemy==2.0.36
pandas==2.2.3
python-dateutil==2.9.0.post0
